requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
Django==5.1.7
django-environ==0.12.0
djangorestframework==3.15.2
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.64
langchain-openai==0.3.19
langchain-text-splitters==0.3.8
langsmith==0.3.45
mypy_extensions==1.1.0
openai==1.84.0
orjson==3.10.18
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.9
pydantic==2.11.5
pydantic_core==2.33.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==78.1.1
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1
zstandard==0.23.0
